Fix buggy line to detect project directory.
Before, we used the process list to find the project
directory. However if the passenger process died, that was
unreliable. A better approach is to just find the first directory in
the common location for putting projects that has a symlink `current`
path - that is likely an active project using passengr.